formmating and add fromJson method
